+++ src/main/java/blackjack/domain/CardNumber.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain;
 
 import java.util.Arrays;
