+++ src/main/java/blackjack/domain/Cards.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain;
 
 import java.util.List;
