+++ src/main/java/blackjack/domain/Dealer.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain;
 
 public class Dealer extends Person {
