+++ src/main/java/blackjack/domain/Person.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain;
 
 public abstract class Person {
