+++ src/test/java/blackjack/CardNumberTest.java
@@ -1,5 +1,6 @@
 package blackjack;
 
+import blackjack.domain.CardNumber;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -8,7 +9,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 class CardNumberTest {
 
-    public static final Stream<Arguments> makeEnum() {
+    public static Stream<Arguments> makeEnum() {
         return Stream.of(Arguments.of(CardNumber.ACE, true), Arguments.of(CardNumber.KING, false));
     }
